Backend/requirements.txt
flask
mysql-connector-python
flask-swagger-ui
marshmallow
python-dotenv
flask-bcrypt
ollama
requests
